+++ CHANGELOG.md
@@ -1,8 +1,11 @@
-## 0.1.15 - UNRELEASED
+## 0.1.15 - 2 Dec 2016
 
 Added some common utilities for reporting progress of long running
 jobs: the `com.walmartlabs.active-status.reporting` namespace, and
 `com.walmartlabs.active-status.component/elapsed-time`.
+
+Added support for creating and reporting status of worker CSPs, in
+`com.walmartlabs.active-status.workers`.
 
 ## 0.1.14 - 18 Nov 2016